backend/requirements.txt
asgiref==3.8.1
astroid==3.2.2
colorama==0.4.6
dill==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
pylint==3.2.4
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-decouple==3.8
sqlparse==0.5.0
tomlkit==0.12.5
tzdata==2024.1